refactor: restructure samples routes and queries
