+++ front/src/index.tsx
@@ -16,8 +16,10 @@ import Login from './pages/login';
 // Measurements
 import MeasurementDetail from './pages/measurements/transfer/detail';
 import MeasurementsList from './pages/measurements/transfer/list';
-import SamplesList from './pages/samples/list';
-import SamplesDetail from './pages/samples/detail';
+
+// Samples
+import WaferList from '@/pages/samples/list/Wafer';
+import WaferDetail from '@/pages/samples/detail/Wafer';
 
 // Events
 import EventDetail from '@/addons/events/pages/EventDetail';
@@ -37,8 +39,8 @@ const routes: RouteType[] = [
   { Component: EventDetail, path: '/event/detail/:id' },
   { Component: MeasurementsList, path: '/measurement/list' },
   { Component: MeasurementDetail, path: '/measurement/detail/:type/:id' },
-  { Component: SamplesList, path: '/sample/list' },
-  { Component: SamplesDetail, path: '/sample/detail/:id' },
+  { Component: WaferList, path: '/sample/list/wafer' },
+  { Component: WaferDetail, path: '/sample/detail/wafer/:id' },
   { Component: Login, path: '/login' },
   { Component: NoMatch, path: '*' },
 ];
